add sign up functionality
