+++ artifact/lib/Screens/sign_up_page.dart
@@ -1,0 +1,149 @@
+import 'package:artifact/Screens/open_page.dart';
+import 'package:artifact/Screens/profile_page.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import "package:flutter/material.dart";
+import "package:artifact/main.dart";
+import 'package:firebase_core/firebase_core.dart';
+
+
+class SignUp_Page extends StatefulWidget {
+  @override
+  _SignUpPageState createState() => _SignUpPageState();
+}
+
+class _SignUpPageState extends State<SignUp_Page> {
+  final emailController = TextEditingController();
+  final passwordController = TextEditingController();
+  @override
+  void dispose() {
+    emailController.dispose();
+    passwordController.dispose();
+
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: Scaffold(
+          body: Column(
+        children: [
+          SizedBox(height: 35),
+          Row(children: [
+            SizedBox(width: 10),
+            IconButton(
+                onPressed: () {
+                  Navigator.pop(context);
+                },
+                icon: Icon(Icons.arrow_back)),
+          ]),
+          SizedBox(height: 65),
+          Image.asset("assets/dsdf1.png",
+              height: 200, width: 200, alignment: topcenter),
+
+          const Padding(
+            padding: EdgeInsets.only(left: 0, top: 0, right: 190, bottom: 0),
+            child: Text(
+              'Username',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+          ),
+
+          Padding(
+            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
+            child: TextField(
+              controller: emailController,
+              decoration: InputDecoration(
+              border: OutlineInputBorder(),
+              hintText: 'Enter your username',
+              ),
+            ),
+          ),
+
+          const Padding(
+            padding: EdgeInsets.only(left: 0, top: 0, right: 200, bottom: 0),
+            child: Text(
+              'Password',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+          ),
+          Padding(
+            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
+            child: TextField(
+              controller: passwordController,
+              obscureText: true,
+              obscuringCharacter: '*',
+              decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: 'Enter your password',
+              ),
+            ),
+          ),
+
+          const Padding(
+            padding: EdgeInsets.only(left: 50, top: 0, right: 180, bottom: 0),
+            child: Text(
+              'Re-enter Password',
+              style: TextStyle(fontWeight: FontWeight.bold),
+            ),
+          ),
+
+          const Padding(
+            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
+            child: TextField(
+              obscureText: true,
+              obscuringCharacter: '*',
+              decoration: InputDecoration(
+                border: OutlineInputBorder(),
+                hintText: 'Re-enter your password',
+              ),
+            ),
+          ),
+
+          const Padding(
+            padding: EdgeInsets.only(left: 80, top: 100, right: 80, bottom: 30),
+            // child: TextButton(
+            //   style: TextButton.styleFrom(
+            //     foregroundColor: Colors.black,
+            //     backgroundColor: Colors.grey,
+            //     textStyle: const TextStyle(fontSize: 20),
+            //   ),
+            //   onPressed: () {},
+            //   child: const Text('Signup'),
+            // ),
+          ),
+
+          TextButton(
+            style: TextButton.styleFrom(
+              foregroundColor: Colors.black,
+              backgroundColor: Color.fromARGB(255, 200, 200, 200),
+              textStyle: TextStyle(fontSize: 16),
+            ),
+            onPressed: signUp,
+            child: const Text('Signup'),
+          ),
+          // new Container(
+          //     child: ElevatedButton(
+          //   child: Text("Signup",
+          //       style: TextStyle(color: Colors.black, fontSize: 16)),
+          //   onPressed: null,
+          // ))
+        ],
+      )),
+    );
+  }
+  Future signUp() async {
+    await FirebaseAuth.instance.createUserWithEmailAndPassword(
+      email: emailController.text.trim(),
+      password: passwordController.text.trim(),
+    );
+    // ignore: use_build_context_synchronously
+    Navigator.push(
+      context,
+      MaterialPageRoute(builder: ((context) {
+        return const MainPage(isLogin: true);
+      }
+    )));
+  }
+}
