+++ artifact/lib/main.dart
@@ -8,6 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:artifact/Screens/open_page.dart';
 import 'package:artifact/Screens/profile_page.dart';
+import 'Screens/sign_up_page.dart';
 import 'admin_login.dart';
 
 const AlignmentGeometry topcenter = Alignment.topCenter;
@@ -22,126 +23,6 @@ Future main() async {
     home: OpenPage(),
   ));
 }
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      home: Scaffold(
-          body: Column(
-        children: [
-          SizedBox(height: 35),
-          Row(children: [
-            SizedBox(width: 10),
-            IconButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                },
-                icon: Icon(Icons.arrow_back)),
-          ]),
-          SizedBox(height: 65),
-          Image.asset("assets/dsdf1.png",
-              height: 200, width: 200, alignment: topcenter),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 0, top: 0, right: 190, bottom: 0),
-            child: Text(
-              'Username',
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
-            child: TextField(
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Enter your username',
-              ),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 0, top: 0, right: 200, bottom: 0),
-            child: Text(
-              'Password',
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
-            child: TextField(
-              obscureText: true,
-              obscuringCharacter: '*',
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Enter your password',
-              ),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 50, top: 0, right: 180, bottom: 0),
-            child: Text(
-              'Re-enter Password',
-              style: TextStyle(fontWeight: FontWeight.bold),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 80, top: 4, right: 80, bottom: 30),
-            child: TextField(
-              obscureText: true,
-              obscuringCharacter: '*',
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-                hintText: 'Re-enter your password',
-              ),
-            ),
-          ),
-
-          const Padding(
-            padding: EdgeInsets.only(left: 80, top: 100, right: 80, bottom: 30),
-            // child: TextButton(
-            //   style: TextButton.styleFrom(
-            //     foregroundColor: Colors.black,
-            //     backgroundColor: Colors.grey,
-            //     textStyle: const TextStyle(fontSize: 20),
-            //   ),
-            //   onPressed: () {},
-            //   child: const Text('Signup'),
-            // ),
-          ),
-
-          TextButton(
-            style: TextButton.styleFrom(
-              foregroundColor: Colors.black,
-              backgroundColor: Color.fromARGB(255, 200, 200, 200),
-              textStyle: TextStyle(fontSize: 16),
-            ),
-            onPressed: () {
-              Navigator.push(context, MaterialPageRoute(builder: ((context) {
-                return ProfileForm();
-              })));
-            },
-            child: const Text('Signup'),
-          ),
-          // new Container(
-          //     child: ElevatedButton(
-          //   child: Text("Signup",
-          //       style: TextStyle(color: Colors.black, fontSize: 16)),
-          //   onPressed: null,
-          // ))
-        ],
-      )),
-    );
-  }
-}
-
 
 class MainPage extends StatelessWidget {
   final bool isLogin;
@@ -158,7 +39,7 @@ class MainPage extends StatelessWidget {
         } else if (isLogin) {
           return LoginPage();
         } else {
-          return SignUpPage();
+          return SignUp_Page();
         }
       },
     ),
